# cellar/gen/http/openapi3.yaml
openapi: 3.0.3
info:
  title: Cellar Service
  description: HTTP service for managing your wine cellar
  version: "1.0"
servers:
- url: http://localhost:8000/cellar
  description: cellar hosts the storage, sommelier and swagger services.
- url: https://goa.design/cellar
  description: cellar hosts the storage, sommelier and swagger services.
paths:
...
  /storage/multi_add:
    post:
      tags:
      - storage
      summary: multi_add storage
      description: Add n number of bottles and return their IDs. This is a multipart
        request and each part has field name 'bottle' and contains the encoded bottle
        info to be added.
      operationId: storage#multi_add
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/AddRequestBody'
              example:
              - composition:
                - percentage: 92
                  varietal: Syrah
                - percentage: 92
                  varietal: Syrah
                description: Red wine blend with an emphasis on the Cabernet Franc
                  grape and including other Bordeaux grape varietals and some Syrah
                name: Blue's Cuvee
                rating: 1
                vintage: 1984
                winery:
                  country: USA
                  name: Longoria
                  region: Central Coast, California
                  url: http://www.longoriawine.com/
              - composition:
                - percentage: 92
                  varietal: Syrah
                - percentage: 92
                  varietal: Syrah
                description: Red wine blend with an emphasis on the Cabernet Franc
                  grape and including other Bordeaux grape varietals and some Syrah
                name: Blue's Cuvee
                rating: 1
                vintage: 1984
                winery:
                  country: USA
                  name: Longoria
                  region: Central Coast, California
                  url: http://www.longoriawine.com/
              - composition:
                - percentage: 92
                  varietal: Syrah
                - percentage: 92
                  varietal: Syrah
                description: Red wine blend with an emphasis on the Cabernet Franc
                  grape and including other Bordeaux grape varietals and some Syrah
                name: Blue's Cuvee
                rating: 1
                vintage: 1984
                winery:
                  country: USA
                  name: Longoria
                  region: Central Coast, California
                  url: http://www.longoriawine.com/
            example:
            - composition:
              - percentage: 92
                varietal: Syrah
              - percentage: 92
                varietal: Syrah
              description: Red wine blend with an emphasis on the Cabernet Franc grape
                and including other Bordeaux grape varietals and some Syrah
              name: Blue's Cuvee
              rating: 1
              vintage: 1984
              winery:
                country: USA
                name: Longoria
                region: Central Coast, California
                url: http://www.longoriawine.com/
            - composition:
              - percentage: 92
                varietal: Syrah
              - percentage: 92
                varietal: Syrah
              description: Red wine blend with an emphasis on the Cabernet Franc grape
                and including other Bordeaux grape varietals and some Syrah
              name: Blue's Cuvee
              rating: 1
              vintage: 1984
              winery:
                country: USA
                name: Longoria
                region: Central Coast, California
                url: http://www.longoriawine.com/
      responses:
        "200":
          description: OK response.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  example: Eius voluptatem.
                example:
                - Laudantium corrupti et.
                - Odit modi.
              example:
              - Ad officia quia.
              - Voluptates sequi nam.
              - Exercitationem nulla mollitia.
              - Itaque iste quibusdam eius officia quaerat dolor.
  /storage/multi_update:
    put:
      tags:
      - storage
      summary: multi_update storage
      description: Update bottles with the given IDs. This is a multipart request
        and each part has field name 'bottle' and contains the encoded bottle info
        to be updated. The IDs in the query parameter is mapped to each part in the
        request.
      operationId: storage#multi_update
      parameters:
      - name: ids
        in: query
        description: IDs of the bottles to be updated
        allowEmptyValue: true
        required: true
        schema:
          type: array
          items:
            type: string
            example: Nihil doloremque magnam qui aut sed.
          description: IDs of the bottles to be updated
          example:
          - Earum hic aperiam quaerat dolorum et.
          - Aspernatur a pariatur nesciunt autem adipisci ut.
          - Porro soluta sapiente ducimus et.
          - Et fugiat.
        example:
        - Dolor mollitia est vel.
        - Dignissimos est non.
        - Ipsam tempora hic qui.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/MultiUpdateRequestBody'
            example:
              bottles:
              - composition:
                - percentage: 92
                  varietal: Syrah
                - percentage: 92
                  varietal: Syrah
                description: Red wine blend with an emphasis on the Cabernet Franc
                  grape and including other Bordeaux grape varietals and some Syrah
                name: Blue's Cuvee
                rating: 1
                vintage: 1984
                winery:
                  country: USA
                  name: Longoria
                  region: Central Coast, California
                  url: http://www.longoriawine.com/
              - composition:
                - percentage: 92
                  varietal: Syrah
                - percentage: 92
                  varietal: Syrah
                description: Red wine blend with an emphasis on the Cabernet Franc
                  grape and including other Bordeaux grape varietals and some Syrah
                name: Blue's Cuvee
                rating: 1
                vintage: 1984
                winery:
                  country: USA
                  name: Longoria
                  region: Central Coast, California
                  url: http://www.longoriawine.com/
              - composition:
                - percentage: 92
                  varietal: Syrah
                - percentage: 92
                  varietal: Syrah
                description: Red wine blend with an emphasis on the Cabernet Franc
                  grape and including other Bordeaux grape varietals and some Syrah
                name: Blue's Cuvee
                rating: 1
                vintage: 1984
                winery:
                  country: USA
                  name: Longoria
                  region: Central Coast, California
                  url: http://www.longoriawine.com/
              - composition:
                - percentage: 92
                  varietal: Syrah
                - percentage: 92
                  varietal: Syrah
                description: Red wine blend with an emphasis on the Cabernet Franc
                  grape and including other Bordeaux grape varietals and some Syrah
                name: Blue's Cuvee
                rating: 1
                vintage: 1984
                winery:
                  country: USA
                  name: Longoria
                  region: Central Coast, California
                  url: http://www.longoriawine.com/
      responses:
        "204":
          description: No Content response.
...
components:
  schemas:
    AddRequestBody:
      type: object
      properties:
        composition:
          type: array
          items:
            $ref: '#/components/schemas/Component'
          description: Composition is the list of grape varietals and associated percentage.
          example:
          - percentage: 92
            varietal: Syrah
          - percentage: 92
            varietal: Syrah
          - percentage: 92
            varietal: Syrah
        description:
          type: string
          description: Description of bottle
          example: Red wine blend with an emphasis on the Cabernet Franc grape and
            including other Bordeaux grape varietals and some Syrah
          maxLength: 2000
        name:
          type: string
          description: Name of bottle
          example: Blue's Cuvee
          maxLength: 100
        rating:
          type: integer
          description: Rating of bottle from 1 (worst) to 5 (best)
          example: 1
          minimum: 1
          maximum: 5
        vintage:
          type: integer
          description: Vintage of bottle
          example: 1980
          minimum: 1900
          maximum: 2020
        winery:
          $ref: '#/components/schemas/Winery'
      example:
        composition:
        - percentage: 92
          varietal: Syrah
        - percentage: 92
          varietal: Syrah
        - percentage: 92
          varietal: Syrah
        - percentage: 92
          varietal: Syrah
        description: Red wine blend with an emphasis on the Cabernet Franc grape and
          including other Bordeaux grape varietals and some Syrah
        name: Blue's Cuvee
        rating: 1
        vintage: 1909
        winery:
          country: USA
          name: Longoria
          region: Central Coast, California
          url: http://www.longoriawine.com/
      required:
      - name
      - winery
      - vintage
    Component:
      type: object
      properties:
        percentage:
          type: integer
          description: Percentage of varietal in wine
          example: 81
          minimum: 1
          maximum: 100
        varietal:
          type: string
          description: Grape varietal
          example: Syrah
          pattern: '[A-Za-z'' ]+'
          maxLength: 100
      example:
        percentage: 75
        varietal: Syrah
      required:
      - varietal
    MultiUpdateRequestBody:
      type: object
      properties:
        bottles:
          type: array
          items:
            $ref: '#/components/schemas/AddRequestBody'
          description: Array of bottle info that matches the ids attribute
          example:
          - composition:
            - percentage: 92
              varietal: Syrah
            - percentage: 92
              varietal: Syrah
            description: Red wine blend with an emphasis on the Cabernet Franc grape
              and including other Bordeaux grape varietals and some Syrah
            name: Blue's Cuvee
            rating: 1
            vintage: 1984
            winery:
              country: USA
              name: Longoria
              region: Central Coast, California
              url: http://www.longoriawine.com/
          - composition:
            - percentage: 92
              varietal: Syrah
            - percentage: 92
              varietal: Syrah
            description: Red wine blend with an emphasis on the Cabernet Franc grape
              and including other Bordeaux grape varietals and some Syrah
            name: Blue's Cuvee
            rating: 1
            vintage: 1984
            winery:
              country: USA
              name: Longoria
              region: Central Coast, California
              url: http://www.longoriawine.com/
          - composition:
            - percentage: 92
              varietal: Syrah
            - percentage: 92
              varietal: Syrah
            description: Red wine blend with an emphasis on the Cabernet Franc grape
              and including other Bordeaux grape varietals and some Syrah
            name: Blue's Cuvee
            rating: 1
            vintage: 1984
            winery:
              country: USA
              name: Longoria
              region: Central Coast, California
              url: http://www.longoriawine.com/
          - composition:
            - percentage: 92
              varietal: Syrah
            - percentage: 92
              varietal: Syrah
            description: Red wine blend with an emphasis on the Cabernet Franc grape
              and including other Bordeaux grape varietals and some Syrah
            name: Blue's Cuvee
            rating: 1
            vintage: 1984
            winery:
              country: USA
              name: Longoria
              region: Central Coast, California
              url: http://www.longoriawine.com/
      example:
        bottles:
        - composition:
          - percentage: 92
            varietal: Syrah
          - percentage: 92
            varietal: Syrah
          description: Red wine blend with an emphasis on the Cabernet Franc grape
            and including other Bordeaux grape varietals and some Syrah
          name: Blue's Cuvee
          rating: 1
          vintage: 1984
          winery:
            country: USA
            name: Longoria
            region: Central Coast, California
            url: http://www.longoriawine.com/
        - composition:
          - percentage: 92
            varietal: Syrah
          - percentage: 92
            varietal: Syrah
          description: Red wine blend with an emphasis on the Cabernet Franc grape
            and including other Bordeaux grape varietals and some Syrah
          name: Blue's Cuvee
          rating: 1
          vintage: 1984
          winery:
            country: USA
            name: Longoria
            region: Central Coast, California
            url: http://www.longoriawine.com/
        - composition:
          - percentage: 92
            varietal: Syrah
          - percentage: 92
            varietal: Syrah
          description: Red wine blend with an emphasis on the Cabernet Franc grape
            and including other Bordeaux grape varietals and some Syrah
          name: Blue's Cuvee
          rating: 1
          vintage: 1984
          winery:
            country: USA
            name: Longoria
            region: Central Coast, California
            url: http://www.longoriawine.com/
      required:
      - bottles
...
    Winery:
      type: object
      properties:
        country:
          type: string
          description: Country of winery
          example: USA
          pattern: '[a-zA-Z ''\.]+'
        name:
          type: string
          description: Name of winery
          example: Longoria
        region:
          type: string
          description: Region of winery
          example: Central Coast, California
          pattern: '[a-zA-Z ''\.]+'
        url:
          type: string
          description: Winery website URL
          example: http://www.longoriawine.com/
          pattern: ^(https?|ftp)://[^\s/$.?#].[^\s]*$
      example:
        country: USA
        name: Longoria
        region: Central Coast, California
        url: http://www.longoriawine.com/
      required:
      - name
      - region
      - country
...